Add StatusMixin class to add status to both sweep measurement
