+++ qcodes/instrument_drivers/Keysight/keysightb1500/KeysightB1500_base.py
@@ -5,7 +5,7 @@ from collections import defaultdict
 
 from qcodes import VisaInstrument, MultiParameter
 from qcodes.instrument_drivers.Keysight.keysightb1500.KeysightB1500_module \
-    import _FMTResponse, fmt_response_base_parser
+    import _FMTResponse, fmt_response_base_parser, StatusMixin
 from qcodes.utils.helpers import create_on_off_val_mapping
 from .KeysightB1530A import B1530A
 from .KeysightB1520A import B1520A
@@ -418,7 +418,7 @@ class KeysightB1500(VisaInstrument):
                                        channels=channels).message)
 
 
-class IVSweepMeasurement(MultiParameter):
+class IVSweepMeasurement(MultiParameter, StatusMixin):
     """
     IV sweep measurement outputs a list of primary and secondary
     parameter.
